+++ src/test.rs
@@ -1165,14 +1165,17 @@ fn fuzzy_inplace_union_test(){
         {
             let non_trusted = &s2 & &s3;
             {
-                let mut s1 = HiSparseBitset::new();
+                let s1 = HiSparseBitset::new();
                 let reference_union: HiSparseBitset = (&s1 | &non_trusted).into();
+
+                let mut s1 = s1.clone();
                 s1.union_with(&non_trusted);
                 assert_equal(&reference_union, &s1);
             }
             {
+                let reference_union = &s1 | &non_trusted;
+
                 let mut s1 = s1.clone();
-                let reference_union: HiSparseBitset = (&s1 | &non_trusted).into();
                 s1.union_with(&non_trusted);
                 assert_equal(&reference_union, &s1);
             }